hl7_pharmacy_transformer/pyproject.toml
[project]
name = "hl7-pharmacy-transformer"
version = "0.1.0"
description = "HL7v2 Pharmacy Transformer - validates MPI messages for Pharmacy system"
requires-python = ">=3.13"

dependencies = [
  "azure-servicebus==7.14.2",
  "azure-identity==1.23.0",
  "hl7apy==1.3.5",
  "message-bus-lib",
  "health-check-lib",
  "processor-manager-lib",
  "event-logger-lib",
]

[dependency-groups]
dev = [
  "ruff==0.11.11",
  "bandit==1.8.6",
  "mypy==1.15.0",
  "uv-secure[faster-async]==0.11.1",
]

[tool.uv.sources]
message-bus-lib = { path = "../shared_libs/message_bus_lib" }
health-check-lib = { path = "../shared_libs/health_check_lib" }
processor-manager-lib = { path = "../shared_libs/processor_manager_lib" }
event-logger-lib = { path = "../shared_libs/event_logger_lib" }

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true
namespace_packages = true
explicit_package_bases = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
  "E",  # Enforces style guide rules (Pycodestyle errors).
  "F",  # Detects various errors in Python code (Pyflakes errors).
  "W",  # Enforces style guide warnings (Pycodestyle warnings).
  "A",  # Detects shadowing of Python built-in functions (Flake8-builtins).
  "PLC",  # Enforces coding conventions (Pylint convention messages).
  "PLE",  # Detects errors in Python code (Pylint error messages).
  "PLW",  # Detects potential issues in Python code (Pylint warning messages).
  "I"  # Sorts and organizes imports (Import-related rules).
]

[tool.bandit.assert_used]
skips = ["*/test_*.py", "*/*_test.py"]  # Stops asserts in tests being flagged.
